requirements.txt
Django~=3.2.11
pymysql==1.0.2
pillow==8.4.0
django-redis==5.2.0
django-redis-cache==3.0.0
djangorestframework==3.13.1
coreapi==2.3.3
